+++ tests/integ/test_dataframe.py
@@ -1326,14 +1326,13 @@ def test_join_on_order(session, local_testing_mode):
     """
     Test that an 'on' clause in a different order from the data frame re-orders the columns correctly.
     """
-    length = 134217728 if not local_testing_mode else None
     df1 = session.create_dataframe([(1, "A", 3)], schema=["A", "B", "C"])
     df2 = session.create_dataframe([(1, "A", 4)], schema=["A", "B", "D"])
 
     df3 = df1.join(df2, on=["B", "A"])
     assert df3.schema == StructType(
         [
-            StructField("B", StringType(length), nullable=False),
+            StructField("B", StringType(), nullable=False),
             StructField("A", LongType(), nullable=False),
             StructField("C", LongType(), nullable=False),
             StructField("D", LongType(), nullable=False),
@@ -2231,8 +2230,8 @@ def test_create_dataframe_large_without_batch_insert(session):
         analyzer.ARRAY_BIND_THRESHOLD = 400_000
         with pytest.raises(SnowparkSQLException) as ex_info:
             session.create_dataframe([1] * 200_001).collect()
-        assert "SQL compilation error" in str(ex_info)
-        assert "maximum number of expressions in a list exceeded" in str(ex_info)
+        assert "SQL compilation error" in str(ex_info.value)
+        assert "maximum number of expressions in a list exceeded" in str(ex_info.value)
     finally:
         analyzer.ARRAY_BIND_THRESHOLD = original_value
 
@@ -2366,7 +2365,7 @@ def test_dataframe_duplicated_column_names(session, local_testing_mode):
             df.create_or_replace_view(
                 Utils.random_name_for_temp_object(TempObjectType.VIEW)
             )
-        assert "duplicate column name 'A'" in str(ex_info)
+        assert "duplicate column name 'A'" in str(ex_info.value)
 
 
 @pytest.mark.skipif(
@@ -2932,7 +2931,7 @@ def test_describe(session):
 
     with pytest.raises(SnowparkSQLException) as ex_info:
         TestData.test_data2(session).describe("c")
-    assert "invalid identifier" in str(ex_info)
+    assert "invalid identifier" in str(ex_info.value)
 
 
 def test_truncate_preserves_schema(session, local_testing_mode):
@@ -3785,7 +3784,7 @@ def test_call_with_statement_params(session):
     # collect
     with pytest.raises(SnowparkSQLException) as exc:
         df.collect(statement_params=statement_params_wrong_date_format.copy())
-    assert "is not recognized" in str(exc)
+    assert "is not recognized" in str(exc.value)
     assert (
         df.collect(statement_params=statement_params_correct_date_format)
         == expected_rows
@@ -3798,7 +3797,7 @@ def test_call_with_statement_params(session):
                 statement_params=statement_params_wrong_date_format.copy()
             )
         )
-    assert "is not recognized" in str(exc)
+    assert "is not recognized" in str(exc.value)
     assert (
         list(
             df.to_local_iterator(
@@ -3811,7 +3810,7 @@ def test_call_with_statement_params(session):
     # to_pandas
     with pytest.raises(SnowparkSQLException) as exc:
         df.to_pandas(statement_params=statement_params_wrong_date_format.copy())
-    assert "is not recognized" in str(exc)
+    assert "is not recognized" in str(exc.value)
     assert_frame_equal(
         df.to_pandas(statement_params=statement_params_correct_date_format.copy()),
         pandas_df,
@@ -3828,7 +3827,7 @@ def test_call_with_statement_params(session):
             ),
             ignore_index=True,
         )
-    assert "is not recognized" in str(exc)
+    assert "is not recognized" in str(exc.value)
     assert_frame_equal(
         pd.concat(
             list(
@@ -3850,7 +3849,7 @@ def test_call_with_statement_params(session):
     # show
     with pytest.raises(SnowparkSQLException) as exc:
         df.show(statement_params=statement_params_wrong_date_format.copy())
-    assert "is not recognized" in str(exc)
+    assert "is not recognized" in str(exc.value)
     df.show(statement_params=statement_params_correct_date_format.copy())
 
     # create_or_replace_view
@@ -3876,7 +3875,7 @@ def test_call_with_statement_params(session):
     # first
     with pytest.raises(SnowparkSQLException) as exc:
         df.first(statement_params=statement_params_wrong_date_format.copy())
-    assert "is not recognized" in str(exc)
+    assert "is not recognized" in str(exc.value)
 
     assert (
         df.first(statement_params=statement_params_correct_date_format.copy())
@@ -3888,7 +3887,7 @@ def test_call_with_statement_params(session):
         df.cache_result(
             statement_params=statement_params_wrong_date_format.copy()
         ).collect()
-    assert "is not recognized" in str(exc)
+    assert "is not recognized" in str(exc.value)
 
     assert (
         df.cache_result(
@@ -3903,7 +3902,7 @@ def test_call_with_statement_params(session):
             weights=[0.5, 0.5],
             statement_params=statement_params_wrong_date_format.copy(),
         )
-    assert "is not recognized" in str(exc)
+    assert "is not recognized" in str(exc.value)
     assert (
         len(
             df.random_split(